test(llm): widen slow-child timeouts and dedup process helpers
Address claude[bot] review on PR #444:
- widen the 50ms timeout to 300ms in both slow-child timeout tests so a
  loaded CI runner cannot kill the fake-codex child before it records its
  PID (the test budget already allows <900ms)
- move shell_single_quote / process_exists / wait_until_process_gone into
  test_support so the codex.rs and mod.rs copies can no longer diverge
